renamed disclaimer tab to info
