+++ ptxboa_streamlit.py
@@ -14,7 +14,7 @@ from app.tab_certification_schemes import content_certification_schemes
 from app.tab_country_fact_sheets import content_country_fact_sheets
 from app.tab_dashboard import content_dashboard
 from app.tab_deep_dive_countries import content_deep_dive_countries
-from app.tab_disclaimer import content_info
+from app.tab_info import content_info
 from app.tab_input_data import content_input_data
 from app.tab_literature import content_literature
 from app.tab_market_scanning import content_market_scanning
